Resolved some install
